+++ examples/data-sources/nxos_port_channel_interface_vrf/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_port_channel_interface_vrf" "example" {
+  interface_id = "po1"
+}