+++ examples/resources/nxos_port_channel_interface_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_port_channel_interface_vrf.example "sys/intf/aggr-[po1]/rtvrfMbr"