+++ examples/resources/nxos_port_channel_interface_vrf/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_port_channel_interface_vrf" "example" {
+  interface_id = "po1"
+  vrf_dn       = "sys/inst-default"
+}